proof: Cogent: update some type system proofs to work with subtyping
Record subtyping defines a record with 'present' fields to be a subtype
of a record with 'taken' fields. This subtyping allows a record with
present fields to be explicitly cast up to a record with the fields missing.

However, if the record field contained a linear value, it would be
silently dropped during up-casting. The subtyping is therefore
restricted to only allow 'droppable' fields, or those with kind {D}.
This restriction was implemented but had been commented out. I have
uncommented the restriction and made the corresponding changes to the
proofs in Cogent.thy.

Most of these changes are quite simple in retrospect. I needed to move
the `specialisation_subtyping` lemma to after the instantiate and
substitutivity lemmas.