proof: subtyping in value semantics
Today I spent most of the day trying to prove the following in the value semantics:
```[] ⊢ t ⊑ t'
  ⟹ Ξ ⊢ v :v t
  ⟹ Ξ ⊢ v :v t'
```
That is, the representation of values does not change from subtype and supertype.
Unfortunately this lemma is not true for three of the value constructors:
1. VRecord: the record representation is an ordered list of fields, but the supertype may use a different ordering;
2. VFunction, VAFunction: the function values contain types which must be changed.

I have started implementing an inductive relation "v_promote" that describes how a value v of subtype t can be promoted to a value v' of supertype t'. This should give us the following lemma:
```value_promotion:
v_promote Ξ v t v' t'
  ⟹ [] ⊢ t ⊑ t'
  ⟹ Ξ ⊢ v :v t
  ⟹ Ξ ⊢ v' :v t'
```
To use this relation from the promotion rule in term semantics v_sem, I needed to thread the \Xi environment through v_sem. I'm not sure whether this is an issue.
When I am next in on Friday, I will continue trying to prove "value_promotion" above, and then try to use it to prove preservation.
Let me know if it looks like I'm going in the wrong direction.

I also noticed that the subtyping relation subty_tcon specifies that all abstract types (TCon) are variant in the type parameters.
I added an assumption for this to the semantics (abs_typing_subty), but I'm not sure whether this is the right assumption to add.
I would have thought that many abstract types would be mutable data structures, and therefore invariant in their type parameters.